編集権限処理、UserIdをDTOに詰めるようにした、その他CSSなど修正

+++ backend/src/main/java/in/tech_camp/prototype_d/entity/UserEntity.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class UserEntity {
-  private Integer id;
+  private Long id;
   private String username;
   private String email;
   private String password;
   private String profile;
-  private Integer affiliationId;
-  private Integer positionId;
+  private Long affiliationId;
+  private Long positionId;
 }
