+++ backend/src/main/java/in/tech_camp/prototype_d/repository/PositionRepository.java
@@ -10,7 +10,7 @@ import in.tech_camp.prototype_d.entity.PositionEntity;
 @Mapper
 public interface PositionRepository {
     @Select("SELECT id FROM positions WHERE position = #{positionName}")
-    Integer findIdByName(String positionName);
+    Long findIdByName(String positionName);
 
     @Insert("INSERT INTO positions (position) VALUES (#{positionName})")
     @Options(useGeneratedKeys = true, keyProperty = "id")
