+++ backend/src/main/java/in/tech_camp/prototype_d/repository/PrototypeRepository.java
@@ -21,7 +21,7 @@ public interface PrototypeRepository {
 
   // 削除権限チェック用
   @Select("SELECT user_id FROM prototypes WHERE id = #{id}")
-  Integer findUserIdById(Long id);
+  Long findUserIdById(Long id);
 
   // プロトタイプ削除
   @Delete("DELETE FROM prototypes WHERE id = #{prototypeId}")
